+++ CHANGELOG.md
@@ -10,6 +10,12 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 - Added support for type hints.
 - `Node` can now be used for generic type hint. Example: `Node[str]`
   represents a node that will have a `str` as result.
+- The class `YapeContext` was added.
+
+### Changed
+- **BREAKING**: `run()` now uses the parameter `context` now and dropped the
+    parameters `ns`, `cached` and `cache_path`. If you used those, you should
+    pass a `YapeContext` object using the equivalent parameters.
 
 
 ## 0.2.1 - 2021-12-02